+++ lib/src/utils/utils.dart
@@ -1,4 +1,3 @@
 export 'generate_secret.dart';
 export 'nip46_encryption.dart';
 export 'nip46_parser.dart';
-export 'no_event_verifier.dart';